+++ services/web/project/resources/users.py
@@ -3,6 +3,7 @@ Director entity resources
 """
 from flask import request
 from flask_restx import Resource
+from flask_login import login_required, current_user
 
 from ..modelschemas.users import UserSchema
 from ..models.users import User
@@ -18,7 +19,7 @@ class UserList(Resource):
     """User Data Resource Class"""
 
     @classmethod
-    def get(cls):
+    def get(cls) -> (dict, int):
         """
         Get method
         :return: error message or json with information about the user
@@ -26,7 +27,7 @@ class UserList(Resource):
         return user_list_schema.dump(User.find_all()), 200
 
     @classmethod
-    def post(cls):
+    def post(cls) -> (dict, int):
         """
         Post method
         :return: json with information about the user
@@ -40,8 +41,9 @@ class UserList(Resource):
 
 class UserListId(Resource):
     """User Data Resource Class"""
+
     @classmethod
-    def get(cls, user_id):
+    def get(cls, user_id) -> (dict, int):
         """
         Get method
         :param user_id: user`s id
@@ -50,39 +52,45 @@ class UserListId(Resource):
         user_data = User.query.get_or_404(user_id)
         if user_data:
             return user_schema.dump(user_data), 200
-        return {'message': USER_NOT_FOUND}, 404
+        return {"status": 404, 'message': USER_NOT_FOUND}
 
     @classmethod
-    def delete(cls, user_id):
+    @login_required
+    def delete(cls, user_id) -> dict:
         """
         Delete method
         :param user_id: user`s id
         :return: error message or successful message
         """
-        user_data = User.query.get_or_404(user_id)
-        if user_data:
-            user_data.delete_from_db()
-            return {'message': "User Deleted successfully"}, 200
-        return {'message': USER_NOT_FOUND}, 404
+        if current_user.is_authenticated and current_user.is_admin:
+            user_data = User.query.get_or_404(user_id)
+            if user_data:
+                user_data.delete_from_db()
+                return {"status": 200, 'message': "User Deleted successfully"}
+            return {"status": 404, 'message': USER_NOT_FOUND}
+        return {"status": 401, "reason": "User is not admin"}
 
     @classmethod
-    def put(cls, user_id):
+    @login_required
+    def put(cls, user_id) -> (dict, int):
         """
         Put method
         :param user_id: user`s id
         :return: json with information about the user
         """
-        user_data = User.query.get_or_404(user_id)
-        user_json = request.get_json()
+        if current_user.is_authenticated and current_user.is_admin:
+            user_data = User.query.get_or_404(user_id)
+            user_json = request.get_json()
 
-        if user_data:
-            user_data.user_login = user_json['user_login']
-            user_data.user_password = user_json['user_password']
-            user_data.first_name = user_json['first_name']
-            user_data.last_name = user_json['last_name']
-            user_data.email = user_json['email']
-        else:
-            user_data = user_schema.load(user_json, session=db.session)
+            if user_data:
+                user_data.user_login = user_json['user_login']
+                user_data.user_password = user_json['user_password']
+                user_data.first_name = user_json['first_name']
+                user_data.last_name = user_json['last_name']
+                user_data.email = user_json['email']
+            else:
+                user_data = user_schema.load(user_json, session=db.session)
 
-        user_data.save_to_db()
-        return user_schema.dump(user_data), 200
+            user_data.save_to_db()
+            return user_schema.dump(user_data), 200
+        return {"status": 401, "reason": "User is not admin"}
